% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shra_bgm.R
\name{bgm_map_object}
\alias{bgm_map_object}
\title{Read BGM with rbgm}
\usage{
bgm_map_object(bgm_file, cum_depth)
}
\arguments{
\item{bgm_file}{path to BGM file}

\item{cum_depth}{vector of cumulative depths}
}
\value{
list of map.vertices and box.data
}
\description{
Read BGM with rbgm
}
\keyword{internal}
