% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{DisplayInitializationExample}
\alias{DisplayInitializationExample}
\title{Spatial Distributions Shiny example}
\usage{
DisplayInitializationExample()
}
\description{
Some examples of the shiny applications based on an incomplete
Atlantis model for Antarctica.
}
