% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initGen.R
\name{get.init.nc}
\alias{get.init.nc}
\title{Function that generates a csv file of NetCDF values}
\usage{
get.init.nc(nc.file, output.file)
}
\arguments{
\item{nc.file}{Name of the NetCDF file containing the initial conditions.}

\item{output.file}{Name of the .csv file where data is written.}
}
\value{
Null (always). Produces a .csv file with the name \code{ouput.file}.
}
\description{
Takes data from a NetCDF file and generates a .csv file containing the box-specific values.
}
\examples{
\dontrun{
nc.file <- "~/Atlantis/RunFiles/SEAP/params/initSEAPaquacult_pH.nc"
output.file <- "oldData.csv" # where to write the data
get.init.nc(nc.file, output.file) # extract the data from NetCDF file
}
}

