% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initGen.R
\name{make.init.csv}
\alias{make.init.csv}
\title{Function that generates two .csv initial condition templates}
\usage{
make.init.csv(grp.file, bgm.file, cum.depths, csv.name)
}
\arguments{
\item{grp.file}{Atlantis group (.bgm) file that defines groups.}

\item{bgm.file}{Box geometry model (.bgm) file used by Atlantis that defines box boundaries and depths.}

\item{cum.depths}{vector of cumulative depths (starting with zero).}

\item{csv.name}{String that is used to identify the two output .csv files.}
}
\value{
Null (always). Produces two .csv files with the names \code{[csv.name]_init.csv} and \code{[csv.name]_horiz.csv}.
}
\description{
Takes data from group file and box geometry file and generates two .csv files that should be copied and modified.
The two files generated are called [csv.name]_init.csv and [csv.name]_horiz.csv.
Also required is the vector of cumulative depths, whos first element is 0.
[csv.name]_init.csv provides all the variables required in the .nc initial conditions file and their default attributes.
[csv.name]_horiz.csv provides box-specific values if the variable is set as customised in the [csv.name]_init file.
See also \code{\link[shinyrAtlantis]{make.init.nc}} for how these files are converted into a NetDF file.
}
\examples{
\dontrun{
grp.file   <- "GBRGroups.csv"
bgm.file   <- "gbr_test.bgm"
cum.depths <- c(0,5,10,20,50,100,200,3000)
csv.name   <- "GBRtemplate"
make.init.csv(grp.file, bgm.file, cum.depths, csv.name)
}
}

