% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinit.R
\name{make.init.object}
\alias{make.init.object}
\title{Collect all data to display}
\usage{
make.init.object(bgm.file, nc.file)
}
\arguments{
\item{bgm.file}{Box Geometry Model (.bgm) file}

\item{nc.file}{NetCDF init file}
}
\value{
list ...
}
\description{
Collect all data to display
}

