% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution Tool v2.R
\name{make.map.objectDistribution}
\alias{make.map.objectDistribution}
\title{Title}
\usage{
make.map.objectDistribution(bgm.file)
}
\arguments{
\item{bgm.file}{}
}
\description{
Title
}
\examples{
## see package readme.md
}

