% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shdist.R
\name{make.sh.dist.object}
\alias{make.sh.dist.object}
\title{Function that generates an object used by sh.dist}
\usage{
make.sh.dist.object(bgm.file)
}
\arguments{
\item{bgm.file}{Box geometry model (.bgm) file used by Atlantis that defines box boundaries and depths.}
}
\value{
R list object.
}
\description{
Takes data from a box geometry model file used by Atlantis to define box boundaries
and uses it to generates a list object that is the parameter to \code{\link[shinyrAtlantis]{sh.dist}}.
}
\examples{
\dontrun{
bgm.file <- "gbr_box_03012012.bgm"
dist.object <- make.sh.dist.object(bgm.file)
sh.dist(dist.object)
}
}
