% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shforce.R
\name{make.sh.forcings.object}
\alias{make.sh.forcings.object}
\title{Function that generates an object used by sh.forcings}
\usage{
make.sh.forcings.object(bgm.file, exchange.file, cum.depth,
  temperature.file = NULL, salinity.file = NULL)
}
\arguments{
\item{bgm.file}{Box geometry model (.bgm) file used by Atlantis that defines box boundaries and depths.}

\item{exchange.file}{NetCDF (.nc) file containing water exchanges between box layers.}

\item{cum.depth}{Vector of depths (starting at 0) that delineate the depths of the water layers for the model.}

\item{temperature.file}{NetCDF (.nc) file containing time-series of temperature values for each water layer in each box. This parameter is not required.}

\item{salinity.file}{NetCDF (.nc) file containing time-series of salinity values for each water layer in each box. This parameter is not required.}
}
\value{
R list object used by \code{\link[shinyrAtlantis]{sh.forcings}}.
}
\description{
Takes data from a box geometry file and a vector of cumulative water layer
depths, as well as netCDF files of: exchanges, salinity and temperature,
and generates a list object that is the parameter to \code{\link[shinyrAtlantis]{sh.forcings}} (see Examples).
}
\examples{
\dontrun{
exchange.file    <- "GBR108_hydro.nc"
salinity.file    <- "GBR108_salt.nc"
temperature.file <- "GBR108_temp.nc"
bgm.file         <- "gbr_box_03012012.bgm"
cum.depth <- c(0,5,10,20,50,100,200,3000) # cumulative water layer depths

input.object <- make.sh.forcings.object(
  bgm.file         = bgm.file,
  exchange.file    = exchange.file,
  cum.depth        = cum.depth,
  temperature.file = temperature.file,
  salinity.file    = salinity.file
)
sh.forcings(input.object)
}
}

