% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinit.R
\name{make.sh.init.object}
\alias{make.sh.init.object}
\title{Function that generates a list object used by sh.init}
\usage{
make.sh.init.object(bgm.file, nc.file)
}
\arguments{
\item{bgm.file}{Box geometry model (.bgm) file used by Atlantis that defines box boundaries and depths.}

\item{nc.file}{NetCDF (.nc) file used by Atlantis to set initial conditions.}
}
\value{
R list object.
}
\description{
Takes data from a box geometry model and a netCDF Atlantis input parameter file and generates a
list object that is the parameter to \code{\link[shinyrAtlantis]{sh.init}} (see Examples).
}
\examples{
\dontrun{
bgm.file <- "VMPA_setas.bgm"
nc.file  <- "INIT_VMPA_Jan2015.nc"
input.object <- make.sh.init.object(bgm.file, nc.file)
sh.init(input.object)
}
}

