% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shprm.R
\name{make.sh.prm.object}
\alias{make.sh.prm.object}
\title{Function that generates a list object used by sh.prm}
\usage{
make.sh.prm.object(bgm.file, grp.file, prm.file)
}
\arguments{
\item{bgm.file}{Box geometry model (.bgm) file used by Atlantis that defines box boundaries and depths.}

\item{grp.file}{Text file (.csv) used by Atlantis containing group attributes.}

\item{prm.file}{Text file (.prm) used by Atlantis containing biological parameters.}
}
\value{
R list object.
}
\description{
Takes data from a box geometry file, a group file, and an Atlantis input parameter file
and generates a list object that is the parameter to \code{\link[shinyrAtlantis]{sh.prm}} (see Examples).
}
\examples{
\dontrun{
bgm.file <- "SEAP_extended_shelf.bgm"
grp.file <- "SEAP_Groups_Aquacult.csv"
prm.file <- "SEAP_biol_pH_Aquacult.prm"
obj <- make.sh.prm.object(bgm.file, grp.file, prm.file)
sh.prm(obj)
}
}

