% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shdist.R
\name{sh.dist}
\alias{sh.dist}
\title{shiny application for generating Atlantis horizontal distributions}
\usage{
sh.dist(map.object)
}
\arguments{
\item{map.object}{}
}
\description{
shiny application for generating Atlantis horizontal distributions
}

