% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shgrowth.R
\name{sh.feeding}
\alias{sh.feeding}
\title{Shiny application for viewing Atlantis feeding and growth data}
\usage{
sh.feeding(grp.file, prm.file, nc.file)
}
\arguments{
\item{grp.file}{Atlantis group file (.csv).}

\item{prm.file}{Atlantis biological parameter file (.prm).}

\item{nc.file}{Initial conditions NetCDF file (.nc).}
}
\value{
Object of class 'shiny.appobj' see \code{\link[shiny]{shinyApp}}.
}
\description{
Takes data from an Atlantis group file (.csv), biology parameter file (.prm),
and initial conditions file (.nc) and produces visualisations of the vertebrate feeding data in the form of a shiny application.
}
\examples{
\dontrun{
bgm.file <- "VMPA_setas.bgm"
nc.file  <- "INIT_VMPA_Jan2015.nc"
sh.init(grp.file, prm.file, nc.file)
}
}

