% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinit.R
\name{sh.init}
\alias{sh.init}
\title{Shiny application for viewing Atlantis initial data}
\usage{
sh.init(input.object)
}
\arguments{
\item{input.object}{from \code{\link{make.map.object}}}
}
\value{
object of class 'shiny.appobj' see \code{\link[shiny]{shinyApp}}
}
\description{
Shiny application for viewing Atlantis initial data
}

