% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyrAtlantis-package.r
\docType{package}
\name{shinyrAtlantis}
\alias{shinyrAtlantis}
\alias{shinyrAtlantis-package}
\title{shinyrAtlantis.
sh.prm   displays data stored in a .prm file
sh.init   displays data stored in the .nc file containing initial values
sh.dist  generates horizontal probability distributions that can be cut-and-pasted into a .prm file}
\description{
shinyrAtlantis.
sh.prm   displays data stored in a .prm file
sh.init   displays data stored in the .nc file containing initial values
sh.dist  generates horizontal probability distributions that can be cut-and-pasted into a .prm file
}

