% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{SpatialDistributionsExample}
\alias{SpatialDistributionsExample}
\title{Spatial Distributions Shiny example}
\usage{
SpatialDistributionsExample()
}
\value{
nothing
}
\description{
Some examples of the shiny applications based on an incomplete
Atlantis model for Antarctica.
}
\examples{
\dontrun{

shinyrAtlantis::SpatialDistributionsExample()
shinyrAtlantis::DisplayParametersExample() # removed as it does not work
shinyrAtlantis::DisplayInitializationExample()
}

}
