% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shdist.R
\name{make.dist.object}
\alias{make.dist.object}
\title{collect data to display the map}
\usage{
make.dist.object(bgm.file)
}
\arguments{
\item{bgm.file}{}
}
\description{
collect data to display the map
}

