% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shdist.R
\name{make.dist.object}
\alias{make.dist.object}
\title{Function that generates an object used by sh.dist}
\usage{
make.dist.object(bgm.file)
}
\arguments{
\item{bgm.file}{Box Geometry Model (.bgm) file used by Atlantis}
}
\value{
R list object
}
\description{
Takes data from a .bgm Atlantis input parameter file and generates a
list object that is the parameter to \code{sh.dist}.
}
\examples{
\dontrun{
prm.object <- make.dist.object(bgm.file)
sh.dist(dist.object)
}
}

