% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinit.R
\name{make.init.object}
\alias{make.init.object}
\title{Function that generates an object used by sh.init}
\usage{
make.init.object(bgm.file, nc.file)
}
\arguments{
\item{bgm.file}{Box Geometry Model (.bgm) file used by Atlantis}

\item{nc.file}{NetCDF initialisation file used by Atlantis}
}
\value{
R list object
}
\description{
Takes data from a .bgm and .nc Atlantis input parameter files and generates a
list object that is the parameter to \code{sh.init}.
}
\examples{
\dontrun{
input.object <- make.init.object(bgm.file, nc.file)
sh.init(input.object)
}
}

