% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shprm.R
\name{make.prm.object}
\alias{make.prm.object}
\title{Make PRM object}
\usage{
make.prm.object(bgm.file, grp.file, prm.file, def.all.file)
}
\arguments{
\item{bgm.file}{BGM}

\item{grp.file}{groups}

\item{prm.file}{parameters}

\item{def.all.file}{defs}
}
\value{
list . . .
}
\description{
Make PRM object
}
\examples{
#See readme
}

