% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinit.R
\name{sh.init}
\alias{sh.init}
\title{Shiny application for viewing Atlantis initialisation data}
\usage{
sh.init(input.object)
}
\arguments{
\item{input.object}{R list object generated from \code{\link[shinyrAtlantis]{make.sh.init.object}}.}
}
\value{
Object of class 'shiny.appobj' see \code{\link[shiny]{shinyApp}}.
}
\description{
Takes data from a .bgm box geometry file and a netCDF Atlantis input parameter file and provides
a visualisation of the data in the form of a shiny application. The two data
files must first be pre-processed by \code{\link[shinyrAtlantis]{make.sh.init.object}},
which generates a list object that is the parameter to \code{sh.init}
(see Examples).

The \emph{Habitat} tab displays the number of water layers and the habitat
cover for each box. Use this tab to check that reef + flat + soft cover sum to 1 for the interior boxes.

The \emph{Abiotic (2D)} and \emph{Abiotic (3D)} tabs display benthic and pelagic abiotic variables.

The \emph{Biotic (2D)} and \emph{Biotic (3D)} tabs display benthic and pelagic biotic variables.

The \emph{Nitrogen} tab displays the initial nitrogen content of biotic groups that are modelled using cohorts.
}
\examples{
\dontrun{
bgm.file <- "VMPA_setas.bgm"
nc.file  <- "INIT_VMPA_Jan2015.nc"
input.object <- make.sh.init.object(bgm.file, nc.file)
sh.init(input.object)
}
}

