% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shprm.R
\name{sh.prm}
\alias{sh.prm}
\title{Shiny application for viewing Atlantis biological parameters}
\usage{
sh.prm(obj)
}
\arguments{
\item{obj}{R list object generated from \code{make.prm.object}}
}
\value{
object of class 'shiny.appobj' see \code{\link[shiny]{shinyApp}}
}
\description{
Takes data from a .bgm, a .csv group file, and .prm Atlantis input parameter file,
and provides
a visualisation of the data in the form of a shiny application. The two data
files must first be pre-processed by \code{make.prm.object}, which generates a
list object that is the parameter to \code{sh.prm}.
}
\examples{
\dontrun{
prm.object <- make.prm.object(bgm.file, grp.file, prm.file)
sh.prm(prm.object)
}
}

