% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shprm.R
\name{sh.prm}
\alias{sh.prm}
\title{Shiny Parameters}
\usage{
sh.prm(obj)
}
\arguments{
\item{obj}{object from \code{\link{mark.prm.object}}}
}
\value{
object of class 'shiny.appobj' see \code{\link[shiny]{shinyApp}}
}
\description{
Shiny Parameters
}

