% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shprm.R
\name{sh.prm}
\alias{sh.prm}
\title{Shiny application for viewing Atlantis biological parameters}
\usage{
sh.prm(obj)
}
\arguments{
\item{obj}{R list object generated from \code{\link[shinyrAtlantis]{make.sh.prm.object}}}
}
\value{
object of class 'shiny.appobj' see \code{\link[shiny]{shinyApp}}
}
\description{
Uses data from an Atlantis box geometry model file, a group file, and a biological parameter file
and provides a visualisation of the data in the form of a shiny application.
The three data files must first be pre-processed by \code{\link[shinyrAtlantis]{make.sh.prm.object}},
which generates a list object that is the parameter to \code{sh.prm} (see Examples).

The \emph{Map} tab displays the spatial distribution of box depths.

The \emph{Groups} tab displays the information stored in the Atlantis group file.

The \emph{Global params} tab shows the global variables that have values provided
in the biological parameters file.

The \emph{Group params} tab allows the user to see a list of possible parameters associated with groups
(Definitions) and the values provided in the biological parameters file (Values).
Use the checkboxes to view the group parameters of interest.

The \emph{Habitat} tab shows which groups are associated with habitats.

The \emph{Distributions} tab shows horizontal and vertical distributions for
chosen groups. Initial values, recruitment and migration parameters can be viewed here.

The \emph{Feeding} tab shows clearance rates, maximum growth rates,
prey availability and refuge-related parameters.

The \emph{Migration} tab provides migrant biomasses (structural and reserve)
and numbers.

The \emph{Recruitment} tab shows the relation between structural and
reserve biomass across all groups.
}
\examples{
\dontrun{
bgm.file <- "SEAP_extended_shelf.bgm"
grp.file <- "SEAP_Groups_Aquacult.csv"
prm.file <- "SEAP_biol_pH_Aquacult.prm"
obj <- make.sh.prm.object(bgm.file, grp.file, prm.file)
sh.prm(obj)
}
}

