% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{SpatialDistributionsExample}
\alias{SpatialDistributionsExample}
\alias{DisplayParametersExample}
\alias{DisplayInitializationExample}
\title{Spatial Distributions Shiny example}
\usage{
SpatialDistributionsExample()

DisplayParametersExample()

DisplayInitializationExample()
}
\value{
nothing
}
\description{
Some examples of the shiny applications based on an incomplete
Atlantis model for Antarctica.
}
\examples{
shinyrAtlantis::SpatialDistributionsExample()

shinyrAtlantis::DisplayInitializationExample()
}
