% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example.r
\name{SpatialDistributionsExample}
\alias{SpatialDistributionsExample}
\title{Spatial Distributions Shiny example}
\usage{
SpatialDistributionsExample()
}
\value{
nothing
}
\description{
Spatial Distributions Shiny example
}
\examples{
shinyrAtlantis::SpatialDistributionsExample()
}

